cache updates
Cache now has 8 byte header:
"ENCORE" and 2 bytes for version number

Cache will be created on startup if it doesn't exist, or remade if the existing cache is invalid.

Cache loading skips over directories that no longer exist, and resizes the songs vector to be the actual loaded number, then checks to see if there's songs that weren't in the cache. If the number of loaded songs differs from the number in the cache, or if new songs were found, it will rewrite the cache with the changed songs.